// front/src/app/models/listing.ts
export class Listing {
    naslov: String;
    grad: String;
    opstina: String;
    mikrolokacija: String;
    ulica: String;
    tip: String;
    kvadrata: Number;
    soba: Number;
    godina: Number;
    stanje: String;
    grejanje: String;
    sprat: Number;
    ukupnoSpratova: Number;
    parking: String;
    mesecneRezije: Number;
    cena: Number;
    opis: String;
    terasa: Boolean;
    lodja: Boolean;
    balkon: Boolean;
    lift: Boolean;
    podrum: Boolean;
    garaza: Boolean;
    basta: Boolean;
    klima: Boolean;
    internet: Boolean;
    interfon: Boolean;
    telefon: Boolean;
    linije: String;
    ime: String;
    prezime: String;
    username: String;
    telefonProdavca: String;
    agencija: String;
    adresaAgencije: String;
    gradAgencije: String;
    telefonAgencije: String;
    pibAgencije: String;
    prodato: Number;
    imagesNames: String;
}